requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
gunicorn==21.2.0
# psycopg2-binary==2.9.9
Flask-Mail==0.9.1
python-dotenv==1.0.0
